Move GuidOrString things to Event.Store.MongoDB - to start making it both v6 and v7 compatible
