+++ Source/Events.Store.MongoDB/Aggregates/AggregateRoot.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
-using Dolittle.Runtime.MongoDB.Serialization;
+using Dolittle.Runtime.Events.Store.MongoDB.Legacy;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
@@ -30,7 +30,7 @@ public class AggregateRoot
     /// <summary>
     /// Gets or sets the id of the Event Source.
     /// </summary>
-    [BsonSerializer(typeof(StringOrGuidSerializer))]
+    [BsonSerializer(typeof(EventSourceAndPartitionSerializer))]
     public string EventSource { get; set; }
 
     /// <summary>
