+++ Source/Events.Store.MongoDB/Events/EventMetadata.cs
@@ -3,7 +3,7 @@
 
 using System;
 using Dolittle.Runtime.Artifacts;
-using Dolittle.Runtime.MongoDB.Serialization;
+using Dolittle.Runtime.Events.Store.MongoDB.Legacy;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
@@ -45,7 +45,7 @@ public class EventMetadata
     /// <summary>
     /// Gets or sets the event source id.
     /// </summary>
-    [BsonSerializer(typeof(StringOrGuidSerializer))]
+    [BsonSerializer(typeof(EventSourceAndPartitionSerializer))]
     public string EventSource { get; set; }
 
     /// <summary>
