+++ Source/Events.Store.MongoDB/Legacy/StringOrGuidFilterDefinition.cs
@@ -7,28 +7,31 @@ using MongoDB.Bson.IO;
 using MongoDB.Bson.Serialization;
 using MongoDB.Driver;
 
-namespace Dolittle.Runtime.MongoDB.Serialization;
+namespace Dolittle.Runtime.Events.Store.MongoDB.Legacy;
 
 /// <summary>
-/// Represents an implementation of <see cref="FilterDefinition{TDocument}"/> for a string or guid.
+/// Represents an implementation of <see cref="FilterDefinition{TDocument}"/> that considers a GUID to be equal to its string representations.
 /// </summary>
 /// <typeparam name="TDocument">The <see cref="Type"/> of the document.</typeparam>
-public class StringOrGuidEq<TDocument> : FilterDefinition<TDocument>
+/// <remarks>
+/// This serializer was introduced to enable backwards-compatibility in the persisted Event Store from v8 to v6 and v7.
+/// </remarks>
+public class StringOrGuidFilterDefinition<TDocument> : FilterDefinition<TDocument>
 {
     readonly FieldDefinition<TDocument, string> _field;
     readonly string _value;
-    
+
     /// <summary>
-    /// Initialize a new instance of the <see cref="StringOrGuidEq{TDocument}"/> class.
+    /// Initialize a new instance of the <see cref="StringOrGuidFilterDefinition{TDocument}"/> class.
     /// </summary>
     /// <param name="field">The field definition.</param>
-    /// <param name="value">The guid or string value.</param>
-    public StringOrGuidEq(FieldDefinition<TDocument, string> field, string value)
+    /// <param name="value">The string value.</param>
+    public StringOrGuidFilterDefinition(FieldDefinition<TDocument, string> field, string value)
     {
         _field = field;
         _value = value;
     }
-    
+
     /// <inheritdoc />
     public override BsonDocument Render(IBsonSerializer<TDocument> documentSerializer, IBsonSerializerRegistry serializerRegistry)
     {
@@ -37,17 +40,26 @@ public class StringOrGuidEq<TDocument> : FilterDefinition<TDocument>
         using var writer = new BsonDocumentWriter(document);
         writer.WriteStartDocument();
         writer.WriteName(renderedFieldDefinition.FieldName);
-        writer.WriteStartDocument();
-        writer.WriteName("$in");
-        writer.WriteStartArray();
-        writer.WriteString(_value);
-        if (Guid.TryParse(_value, out var guidValue))
-        {
-            writer.WriteBinaryData(new BsonBinaryData(guidValue, GuidRepresentation.Standard));
-        }
-        writer.WriteEndArray();
-        writer.WriteEndDocument();
+        RenderEqualityQuery(writer);
         writer.WriteEndDocument();
         return document;
+    }
+
+    void RenderEqualityQuery(IBsonWriter writer)
+    {
+        if (Guid.TryParse(_value, out var guidValue))
+        {
+            writer.WriteStartDocument();
+            writer.WriteName("$in");
+            writer.WriteStartArray();
+            writer.WriteString(_value);
+            writer.WriteBinaryData(new BsonBinaryData(guidValue, GuidRepresentation.Standard));
+            writer.WriteEndArray();
+            writer.WriteEndDocument();
+        }
+        else
+        {
+            writer.WriteString(_value);
+        }
     }
 }
